mypy: fix no-any-return/assignment/arg-type/misc errors across 12 files (17 -> 0); rename shadowed loop var in _target_distribution_analyzer_features.py

+++ src/mlframe/training/phases.py
@@ -106,7 +106,7 @@ def _try_get_rss_gb() -> float:
     """Best-effort current process RSS in GB; 0.0 if psutil missing."""
     try:
         import psutil
-        return psutil.Process().memory_info().rss / (1024 ** 3)
+        return float(psutil.Process().memory_info().rss / (1024 ** 3))
     except Exception:
         return 0.0
